refactor: move directory pages component
