+++ src/Router.tsx
@@ -4,8 +4,8 @@ import Main from './layouts/Main/Main';
 
 import HomePage from './pages/HomePage';
 import SearchPage from './pages/SearchPage';
-import NotFoundPage from './pages/NotFoundPage/NotFoundPage';
-import UserPage from './pages/UserPage/UserPage';
+import NotFoundPage from './pages/NotFoundPage/';
+import UserPage from './pages/UserPage/';
 import BookmarkPage from './pages/BookmarkPage';
 
 export default function Router() {
