+++ src/pages/NotFoundPage/index.ts
@@ -1,0 +1,3 @@
+import NotFoundPage from './NotFoundPage';
+
+export default NotFoundPage;
