+++ src/pages/UserPage/index.ts
@@ -1,0 +1,3 @@
+import UserPage from './UserPage';
+
+export default UserPage;
